prov/efa : fix a bug in cuda long message protocol
Need to call rxr_inline_mr_reg() in rxr_pkt_handle_long_rtm_sent()
even when we are not do inline_mr_reg(), because that function
do two things:

   It will set tx_entry->iov_index and tx_entry->iov_offset

   It will register memory for tx_entry->iov.

If tx_entry->desc is provided, we do not need to register memory for
tx_entry->iov, but we still need to change the value of tx_entry->iov_index.

This patch rename the function rxr_inline_mr_reg() to rxr_prepare_mr_send(),
and do not do registration if tx_entry->desc[i] is provided.

Signed-off-by: Wei Zhang <wzam@amazon.com>