prov/efa: do not unset RDMAV_HUGEPAGES_SAFE if user has it
Currently, the efa provider unconditionally set and usnet the
environment variable RDMAV_HUGEPAGES_SAFE. Even if user
has the environment variable, it will be unset by the proivder.

This patch address the issue by introducing a global variable:

    efa_set_rdmav_hugepages_safe

This variable record whether the provider set the environment
variable, and is used to determine if the proivder should unset.

Signed-off-by: Wei Zhang <wzam@amazon.com>