[Experimental]: launch a cuda kernel to copy data from host to GPU
This patch demonstrate an experimental apporach to address

    https://t.corp.amazon.com/V202523328

which is GDR stack hang with NCCL when NCCL is using the default
stream.

The problem is during message transfer, libfabric need to copy data from
host to GPU. However, when default stream is used, and there are kernel
and cudaMemcpy on it, which blockes any regular data copy methods such
as cudaMemcopy, cudaMemcopyAsync, cuMemcpyHtoDAsync.

This PR would do the following:

1. During initialization, register the packet pool memory with CUDA
   using cudaHostReigster. Doing this allow CUDA kernel to access the
   memory.

2. During initialization, create a stream

3. During message transfer, instead of using cudaMemcpy(), launch a kernel
   to copy data from host to device.

Note this patch introduced some cuda specific changes to libfabric, including
the change to Makefile.am, add a script to generate libtool object. It is
not clear whether that libfaric community will accept it. It is provided
mainly to demonstrate the related work for this approach to work.

Signed-off-by: Wei Zhang <wzam@amazon.com>