[debug] prov/efa : skip creating qp for btl ofi
We noticed that when btl/ofi is used, alltoall performance is degraded.
Even though the btl/ofi component was not used.

This commit demonstrate that by skipping qp creation for btl ofi
component. The way we distinguish btl/ofi and mtl/ofi is from the
requirement of delivery complete of the btl/ofi component.

Signed-off-by: Wei Zhang <wzam@amazon.com>