fabtests: add functional/rdm_multi_client test.
This program tests the functionality of RDM endpoint in the case
that a persistent server does ping-pong with multiple clients that
come and leave in sequence. The client connects to a server, sends
ping-pong, disconnects with the server by cleaning all fabric
resources, and repeats.
It will re-use the first client's address for the subsequent clients
if the `fi_setname` API is implemented for the tested provider.

Signed-off-by: Shi Jin <sjina@amazon.com>