prov/efa: add tx_entry_list and rx_entry_list to peer
Currently, the struct efa_rdm_peer has a member use_cnt,
which is the total number tx_entry and rx_entry associated
with the peer.

This patch used tx_entry_list and rx_entry_list to replace
use_cnt, so that when removing a peer, all the tx_entry/rx_entry
associated with it can be removed.

To facilitate this change, this patch did some refactor:

1. a new function efa_rdm_peer_init() was extracted from
   efa_conn_rdm_init(), and the initialization of
   peer->tx_entry_list and peer->rx_entry_list was added
   to it.

2. the function efa_rdm_peer_reset() was renamed
   efa_rdm_peer_clear(), and the release of peer->tx_entry_list
   and peer->rx_entry_list was added to it

Signed-off-by: Wei Zhang <wzam@amazon.com>