prov/efa: refactor error handling and fix a bug
Currently, if application want to receive RNR error completion
entry, any packet encountered RNR will caused an error completion
to be written. This is wrong because error completion should only
be written for a send opeartion. However, not all packets are
assoicated with send completion. For example, EOR and CTS packets
are associated with a receive completion, thus writing RNR error
completion for when these packets encountered RNR is wrong.

To addres this issue, this patch did a series of refactors:

1. the function rxr_cq_handle_error was splitted into two
   functions rxr_pkt_handle_send_error() and
   rxr_pkt_handle_recv_error().

2. rxr_pkt_handle_send_error() will handle RNR differently
   according to pkt_entry's x_entry type. It only write
   RNR error completion if x_entry is tx_entry.

3. rxr_cq_handle_tx_error() was renamed to rxr_cq_write_tx_error()
   and return type changed to void.

4. rxr_cq_handle_rx_error() was renamed to rxr_cq_write_rx_error()
   and return type changed to void.

5. rxr_read_handle_error() was renamed to rxr_read_write_error()
   and return type changed to void.

Signed-off-by: Wei Zhang <wzam@amazon.com>