prov/efa: fix a bug that handshake encoutered RNR was dropped
Currently, if handshake packet encountered send error completion
, the error completion is ignored. This has caused
application to hang because peers need handshake packets to proceed
with communication.

This patch address the issue by handling different error differently:

If the error because of RNR, the peer is added to handshake_queued_peer_list,
so the handshake will be sent again.

If the error because of remote EP has been teared down, the error
is ignored.

For other error, an eq entry is written.

Signed-off-by: Wei Zhang <wzam@amazon.com>