prov/efa: fix a bug in rxr_cq_handle_error()
Currently, if application want to receive RNR error completion
entry, any RNR packet error will caused an RNR error completion
to be written, even if the packet is not associated with a send
operation. For example, EOR and CTS packets associated with
receive operations, and HANDSHAKE packet is not associated with
any user operation.

This is wrong because error completion should only be written for
a send opeartion.

To addres this issue, and to make the code more clear,
this patch did a series of refactors:

    1. the function rxr_cq_handle_error() was splitted into two
       functions rxr_pkt_handle_send_error() and
       rxr_pkt_handle_recv_error().

    2. rxr_pkt_handle_send_error() handle RNR differently
       according to pkt_entry's x_entry type. It only write
       RNR error completion if x_entry is tx_entry.

    3. All the places, rxr_cq_handle_error() was called to handle
       a general error was replaced with a call to efa_eq_write_error(),
       which is called by rxr_cq_handle_error() previously.

Signed-off-by: Wei Zhang <wzam@amazon.com>